+++ miditok/tokenizations/mmm.py
@@ -320,7 +320,7 @@ class MMM(MIDITokenizer):
             elif tok_type == "Tempo":
                 # If the tokenizer includes tempo tokens, each Position token should be followed by
                 # a tempo token, but if it is not the case this method will skip this step
-                tempo = int(token.split("_")[1])
+                tempo = float(token.split("_")[1])
                 if tempo != tempo_changes[-1].tempo:
                     tempo_changes.append(TempoChange(tempo, current_tick))
             elif tok_type == "TimeSig":
