+++ miditok/tokenizations/remi.py
@@ -256,7 +256,7 @@ class REMI(MIDITokenizer):
             elif token.split("_")[0] == "Tempo":
                 # If your encoding include tempo tokens, each Position token should be followed by
                 # a tempo token, but if it is not the case this method will skip this step
-                tempo = int(token.split("_")[1])
+                tempo = float(token.split("_")[1])
                 if tempo != tempo_changes[-1].tempo:
                     tempo_changes.append(TempoChange(tempo, current_tick))
             elif token.split("_")[0] == "Pitch":
